Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/406e465f-9d2a-4940-813c-d567310fc199). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I066d9c1c5f0b3ed0e2b5731db916414bb4137cf6