Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ce4b5dd-a32a-42c7-8784-73adc4b1721b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46afbaa9d8ac4e8aa395a28b3d58877c253bfc5f