Revert "Implement clamp-growth-limit for userfaultfd GC"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/425c6139-0fa9-41b5-833c-c1a99bb5128a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d747da5e77dafb4f6ec816fa3f7ecaa67231fd4