Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/32a87927-3788-4959-9a88-a2babe5227f9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98ecd22a95113e2d8908484b21349648032d19b9