Revert "Revert^3 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffda9fa6-1bad-4f3d-8c4a-77b94a41eb04). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcdd7b03cfb965a7515f70bed9267f744d6f80d3