Use kLargeObjectAlignment for FreeListSpace
This patch is part of the chain preparing for making kPageSize
non-constexpr in a future patch.

Replace FreeListSpace::kAlignment with kLargeObjectAlignment.

The replacement is not required and is accomplished to make it
more obvious which value is used for alignment in the space.

Without this change, both kLargeObjectAlignment and
FreeListSpace::kAlignment (always equal to kLargeObjectAlignment)
are used in large_object_space.cc, causing unnecessary duplication
in ways to refer to the same value:
 - direct reference to kLargeObjectAlignment;
 - indirect reference via kAlignment defined as kLargeObjectAlignment.

Test: Same as for I5430741a8494b340ed7fd2d8692c41a59ad9c530.
      The whole patches chain was tested as a whole.
Change-Id: I67129cf6cd6168c2d10adc7b2ebd7f5ab7109ede