Refactor OatQuickMethodHeader for assembly stubs.
Refactor OatQuickMethodHeader to free up the most significant bit.
Use very large magic value for the header and write the stub code
size just before the header instead.

This should allow us to have oat files close to 4GB in size.

It should also make it easier to DCHECK that we see stub header
when we expect it in other parts of the code (e.g. stack walk).

Bug: 288215378
Bug: 288833213
Bug: 297093110
Test: ./art/test.py -b --host
Change-Id: I3799c23b09d827d13be2b28a22629de1545f5c70