test: Run-time detect target features by default
If neither --instruction-set-variant nor --instruction-set-features are
provided, ART will generate generic code. In tests we do not pass these
arguments by default, meaning that codegen which depends on specific
features being enabled is not tested by default.

This patch sets the default --instruction-set-features argument for
tests to be 'runtime', so that tests will generate code based on the
features supported by each test device.

Runtime ISA feature detection is currently supported on Arm64. If a
target does not support run-time detection, this change has no effect
and the selected features will remain as the default configuration.

Tested on oriole, shiba.

Test: art/test.py --target --optimizing
Test: art/test.py --target --optimizing
      --run-test-option='--instruction-set-features "default"'
Test: art/test.py --target --optimizing
      --run-test-option='--instruction-set-features ""'

Change-Id: Ic3311bfe7566bb09e6f915e988d629626f8331cb