Revert "Refactor OatQuickMethodHeader for assembly stubs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd97bc75-ca98-451c-934a-b9b4987d5f82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25f5a729906dbdd3657402ae9743789004811198