Revert "Refactor OatQuickMethodHeader for assembly stubs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f941722-24ab-4730-9e46-e263a47f9632). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I44903c4c80743cd08db497fd3790234071867250