Revert "Refactor OatQuickMethodHeader for assembly stubs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c2d5a25-de37-4893-878a-3f349030eb43). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie6ea9a7f1c9f88f5ec1732bd0e98108520d248ce