Revert "Refactor OatQuickMethodHeader for assembly stubs."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/042c4b67-a041-4394-afb5-705008bdca81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c04939ef7254cb9d254700a1da50a8bc8764ae5