riscv64: [codegen] Add VisitNullCheck
Also enable codegen for NullCheck and fix related entrypoint
issues. For now, this is using explicit null checks.

Change the default value of the implicit null checks flag in
the `CompilerOptions`. In `dex2oat` we expect the default
value to be false and override it to true when needed by the
selected architecture. This aligns with the behaviour in
`Runtime` which is the source of this information for JIT.
We do not change the default value for implicit stack
overflow checks flag yet because it requires additional
adjustments to avoid breaking certain gtests.

Test: aosp_cf_riscv64_phone-userdebug boots.
Test: run-gtests.sh
      # Ignore pre-existing timeout in `TestImageLayout`.
Test: testrunner.py --target --64 --optimizing
      # Ignore 49 pre-existing failures.
Bug: 283082089
Change-Id: If663d3279da5e6c53669860cefa7185c53e7e146