Change how the test checks the artifact timestamps
CompOsDenialHostTest checks that CompOS artifact isn't used in the
error cases, and Android uses the compilation that happens during
early boot. The test used to check this by comparing timestamp:

  time of CompOS compilation < time of early boot compilation

Appeartly this isn't necessarily true on some device implementation,
in which the device boots from timestamp 0.

This change modifies the test to compare the file names and timestamps
of the CompOS compilation and after the reboot. In the denial cases,
both must be different. Although this can be more fragile in the future
since any change would still pass the test in an unexpected way. A
regular expression was added to mitigate this.

As another mitigation, and also a test coverage improvement, is to let
CompOsSigningHostTest check the file name and timestamp snapshot. This
is opposite: the timestamps have to match to pass the tests. As a
result, previous time checks are less effective and now removed.

Bug: 270528861
Test: atest odsign_e2e_tests_full
Change-Id: I7e367266c6c14f2ec4c8237c710c6d1eedc2623a