riscv64: check_cfi sees trailing unimp
The nterp_deliver_pending_exception trampoline has
a deliberate `unimp` after the jump. The jump is
phrased as a `call` to save the RA but a return is
not expected and the `unimp` would catch a bad
return.

The `check_cfi` tool performs a disassembly to
examine each PC of each function. Prior to this
patch, objdump will simply ignore the final
`unimp` because it looks like the past of a
function. Adding the `--disassemble-zeroes` flag
forces the last `unimp` to be observable and hence
analyzed by the tool.

Test: m check_cfi
Change-Id: Ieb03a7874c521fc49ecf2541167d623929eafe28