Revert "Prevent repeatedly generating runtime images."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aefc7982-77ae-412c-85dc-d596aceab648). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic6c284bd5d9c4501b2a9d81ca6743adda980a174