Revert "Prevent repeatedly generating runtime images."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb448bcc-2fd1-473f-ace0-201b6707a5df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9be86f491fac348bfad2b96f38a3899157c40fab