Should not check if product is treblelized from vndk version
There are some logic in libnativeloader to check if product is
treblelized by checking ro.product.vndk.version. However, this property
is no longer valid to check if product is treblelized, because of vndk
deprecation. Also product is force-treblelized if the device is
treblelized from R, so no extra logic is required to check this.

Bug: 290159430
Test: Cuttlefish build and run succeeded
Change-Id: Ieee3ff2dde4244e7c46420ad8dde5f8e5cf249e4