Fix 909-attach-agent to work on target.
Clean up an old kludge that tried to find the agent lib in
java.library.path and copy it to the test directory. That hasn't been
working since linker namespaces and NATIVELOADER_DEFAULT_NAMESPACE_LIBS
were introduced, but the test has been disabled so it wasn't noticed.
Even if the copy worked the library wouldn't be a directory that the
linker config recognised, so it wouldn't find dependencies like
libbase.so.

Test: art/test/testrunner/testrunner.py --target --32 909
Test: art/test/testrunner/testrunner.py --host 909
Bug: 300654679
Change-Id: Id0b3a32ca757af63e878d3c5b0ae70919685cb65