Initialize gPageSize at runtime via sysconf
This changes gPageSize to a global constant const in page size agnostic
configuration, dynamically initialized with the runtime-determined page
size value. This finishes adding basic support of page size agnostic
ART configuration, however without yet enabling it.

With page size agnosticism disabled, gPageSize etc. derived values in
the global scope are still constexpr.

As part of that, introduce helpers for gPageSize and the derived
constants in the global scope, in a way that guarantees correct static
initialization order:
 - GlobalConst is a helper class that acts as a global constant;
 - ART_PAGE_SIZE_AGNOSTIC_DECLARE etc. - helper macros for declaring the
   constants either as the global constants or constexpr depends on the
   configuration.

The helpers are used for the gPageSize and derived values stored in the
global scope.

Test: Same as for I5430741a8494b340ed7fd2d8692c41a59ad9c530.
      The whole patches chain was tested as a whole.
Change-Id: Id1c18004346ba5c6c94e02cdf8b0b0bb3b99af70