Revert "Fix memory leak in VMRuntime_setDisabledCompatChangesNative."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ff0b26c-773f-43f7-bc93-dd872b29bf68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb5d991e1ffad14dae3ed04a2bd72b673fd98c51