Revert "Add more checks for helping with b/284236137."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed12e877-cf69-4883-9d5a-b6779fc6756d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d7f6b8daa3afd60280af94672d38a256c96d12f