Revert "riscv64: [codegen] Rewrite and enable Select."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20a0ab28-731f-4ff7-be74-bea0c7bbdf0f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dde7e6cb10c77debf1527c9b00515b16e743105