Revert "riscv64: [codegen] Rewrite and enable Select."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eec59d36-88f1-4232-9fa5-df8f325bf4af). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I245bc9d4bd7734b63465981953e59cf7ba5b7acb