Ignore read-barrier invocations on immune-space objects
We need to gray the immune objects on private-dirty pages only as
others can't have references to moving-space. And these private-dirty
immune objects are handled by the gc-thread at thread-flip. Therefore,
the entire infrastructure for dealing with read-barriers on immune-space
references is not required.

This may help save us from unecessarily dirtying some immune-space pages
due to this.

Test: art/test/testrunner/testrunner.py
Change-Id: I9e14681a22a95c1fa632732cb80ccaf7923d5b44