Revert "Add branch profiling in baseline compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f5787581-7f97-45ac-bc1a-d14425689c7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idad5efa858be7b1babc2a4bb7aac4082c85fd4e4