Revert "Add branch profiling in baseline compiler."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b1a39c63-27e2-40f1-aec1-e7c81ab85008). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48dcc555a70670e2fe78fd93d55d06f4e7aee59f