[NOT FOR COMMIT] squash of 4 CLs
Improve linear loop optimization overflow checks

We can move the logic from GenerateLastValueLinear to GenerateCode,
where we can:
 * be more granular, and
 * reuse that code for other methods.

To do so we add an allow_potential_overflow variable. This is done
because BCE does perform operations that might overflow, but it
uses HDeoptimize instructions to guard against that. Loop
optimization doesn't add HDeoptimize instructions so we must be
more careful.

Bug: 231415860
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b

Improve linear induction var range creation

We can now detect and remove loops that require an is_taken test e.g.

int a = 0;
for (int i = 0; i < n; i++) {
  a += 1;
}

can be turned into `if (n < 0) then 0 else n`.

Part of this logic can be reused in the future to help eliminate
BoundsCheck instructions.

Bug: 304967775
Fixes: 304967775
Test: art/test/testrunner/testrunner.py --host --64 -b --optimizing

Make periodic loop optimization safe from overflow

The trip number is calculated by doing:
    (end - start) + (step - 1) / step
(Note that we add step - 1 as a way of doing ceiling).
This way of calculating can overflow and produce the wrong result
if e.g. end and start are in opposite sides of the spectrum.

We can add checks to make sure that the calculation doesn't overflow.

Bug: 256858062
Fixes: 256858062
Test: art/test/testrunner/testrunner.py --host --64 -b --optimizing

Align induction var range checks and uses

We were requiring some conditions but we weren't checking all of them.

Test: art/test/testrunner/testrunner.py --host --64 -b --optimizing
Change-Id: I7918f97030a25749a2bdeeb7416fd5fe1865b873