Revert "Fix performance-unnecessary-value-param clang-tidy issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bcc5cc95-e354-467a-a5d7-27e6f97fa6c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib837e9cdedbee87951b2abaef0973ccd61617be9