Revert "Fix modernize-use-nullptr clang-tidy issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bcc5cc95-e354-467a-a5d7-27e6f97fa6c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb115d004a10876e2f212d76198851e22a6ac985