Revert "Fix bugprone-argument-comment clang-tidy issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bcc5cc95-e354-467a-a5d7-27e6f97fa6c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic56f2ed69cac80f409d34d4a714873bf8db1b3fb