Revert "Remove old and duplicated logic in picking up the best a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b624aa13-c57b-4098-b0ab-276fb12f67f4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46b4669457da8a46256e268f5e61313fd15af7f7