Revert "Reland "Add trace points for class loading.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fcb5c071-2c5a-4056-9cf2-375517443802). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d84b77a181282a07c87fe20ca864d83ebd1eaf2