[DON'T MERGE] [RFC] Optimize access to GlobalConst
The change ensures the global constants can be accessed without
conditional on whether they were initialized. Before this the constants
were accessed via function that would check if the initialization was
complete. If it wasn't, then the initialization was performed before
the function returns. With this change, the initialization happens
early when the DSO (or static binary) is loaded into the process - via
the init_array. This way the conditional can be removed everywhere the
constants are accessed during normal execution.

Additionally to solving the issue with initialization order across
libraries, the change ensures the global constants can be accessed
directly via offset instead of being accessed via the Global Offset
Table.

This patch is an RFC on the alternative approach to ensure static
initialization order is correct for the gPageSize and dependent
global constants however possibly it can be seen as providing an
infrastructure for all global constant values independently of page
size agnosticism support.

Change-Id: I730c81a0138b5ee92a2ac30168e3ba7712aa2820