Revert "Update art test expectations as part of R8 roll"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/47c23b50-204b-48d7-9ac5-8600d3314d68). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99a2b724b3f500f345522b30331fa28ec693f39f