Store gPageSize as a log2 - gPageSizeLog2
When the gPageSize was a constexpr, the compiler had information
that it is a power of two, and could replace multiplication/division
with logical shifts. However, once it became a dynamically initialized
value, the information was lost.

Represent gPageSize with its log2 to preserve the information for
possible compiler optimizations.

Some of the effects of this change are as follows:
 - where there is multiplication/division by a page size, these
   operations can be emitted as logical shifts which is expected
   to be more efficient;
 - where the page size value is used for other purposes, an extra
   logical shift would appear.

With caching of the value (by compiler or manually) the overhead
from additional shift should be negligible, however benefits from
replacing, for example, division with logical shift are expected
to be significant.

The tests were run for legacy 4K, page size agnostic 4K and 16K.

Test: art/tools/run-gtests.sh
Test: art/test/testrunner/testrunner.py --target --64
Test: art/tools/run-libcore-tests.sh --mode=device --variant=X64
Test: art/tools/run-libjdwp-tests.sh --mode=device --variant=X64
Change-Id: I1f189404933c6e3d9a286c835b220726128a2ae7