Align BitMemoryRegion down to prior kMaxPageSize
The page size can now be a static global non-constexpr value, which
means that references to this value can incur an additional overhead
that is especially observed during exception throwing.

In this change, avoid this potential overhead and thus optimize
exception throws by replacing BitMemoryRegion's usage of the static
global gPageSize value with the use of kMaxPageSize instead, which is a
constexpr in all configurations and thus incurs zero additional overhead
to access.

The context for this change is that previously a change was made that
set each BitMemoryRegion's data pointer to the immediately preceeding
page boundary, thereby optimizing performance when accessing addresses
offset from this pointer.

By instead aligning down to the immediately preeceding kMaxPageSize
boundary, the resulting pointer is still page-aligned, and the optimized
performance of subsequent accesses should be maintained while also
avoiding any overhead from accessing the static global.

Test: art/tools/run-gtests.sh
Test: art/test/testrunner/testrunner.py --target --64
Test: art/tools/run-libcore-tests.sh --mode=device --variant=X64
Test: art/tools/run-libjdwp-tests.sh --mode=device --variant=X64

Change-Id: I9ee90e41e22a63e3693356f79cc8df90d03d6b20