Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05999607-03fd-4393-993f-51ddbd7b29c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I70004805f0b64450844fd38af6773bc144f56a3a