Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5b2d9fa4-879d-46c2-bb96-2625eb6c1150). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6bc958166c2730c940464a880b5ac2b2bfb6de91