Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ee3b94ac-07bb-4f4c-b21a-3c6e010994b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf60810a31730485290bc70bf949d78f7d4e8c94