Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ea3a071-01ed-4f22-9b42-e23dc0f973dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6bad5993c42a33e82a0b743b130d1c249f0a9a93