Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ebcaca4f-8210-40d1-8336-635ae1222a3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00744273b3a893c8988ca3341b9b2b5aed11c9f6