Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aba71f35-1939-46a0-89a7-87399ffdc421). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieacb056d8eccca0ae628cebbf0335794c9ae4ad2