Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/118775f4-bb30-40b2-8629-8ecc132a62a3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8541cc59ae6d353a35ffe6a2c618b9af210bddfd