Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fbd6c402-4329-4371-b56d-f6835e949978). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0851fc70aeadf3f73b5663605af1ad609a0ed900