Change `dexoptPackages` to support multiple passes.
This CL changes `dexoptPackages` so that it can report progress and
results for multiple passes. It enables us to add one more pass for
background dexopt and collect stats of the new pass.

This is generally a no-op change, except for that it changes the outputs
of the `pm art dexopt-packages` command.

Bug: 242170869
Test: atest ArtServiceTests
Test: adb shell pm art dexopt-packages -r bg-dexopt
Test: adb shell pm art dexopt-packages -r first-boot
Change-Id: I6197ec92caca0e30a4efbc394faf24c778ba0b8a