Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ec2849d-b9ee-409e-a5fb-541387a98801). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a8c5c84c70c097b6d96c10eadccee766786ca27