Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a79a56bb-fbe6-4c21-9a32-0763c6c58483). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I353e1766f9f37cad1af4839fa50c4c0c4af3d171