Add a new profman flag --force-merge-and-analyze.
When the flag is set, profman merges profiles even if the difference
between before and after the merge doesn't meet the thresholds, and
returns `kCompile` if there is any difference or
`kSkipCompilationSmallDelta` if there is no difference.

The behavior of this flag is the same as --force-merge, except for that
it returns different codes to indicate whether there is any difference
or not. We can't change the return code of --force-merge because
installd on S and T relies on it to be `kSuccess`.

Bug: 242170869
Test: m test-art-host-gtest-art_profman_tests
Change-Id: I59fdc64a66284ec66c4b7a7c355df9e902dfee00