Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/447ace1e-09c8-40d6-9727-f4947d29298a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85a4cb1b4db52f2542c3a6ac1dbc3862c5d63f7a