Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3084a844-31b3-428a-aeec-796a50b5e01d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id90a46428f3a896ff4fe3da2d9ca2e53c9a7fade