Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/15204118-03c5-478a-80cb-3b8b6a1e0c41). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I71c1552c7b33fced0544de2c5827b5508456a093