Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27de7a4a-a63c-4a50-b06e-008a68892e0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia30b723b66ecd449ae14378467ee9fc7077a5702