Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/321c4127-9f7c-4e00-99fc-8ee85e968cbd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If757d083bddb6fe5af6f075e3ab1554eacad44cd