Revert "Add a new profman flag --force-merge-and-analyze."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ae9634a-1d2e-49f7-8630-183885272221). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9218a5e956e1a2edb6d2f2c980a1395d57f783d5