Revert "Change `dexoptPackages` to support multiple passes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ae9634a-1d2e-49f7-8630-183885272221). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b8ed6496fd8e63790bd3f3bfb20b07028561b2d