Revert "Change `dexoptPackages` to support multiple passes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc352553-0fde-4a1b-949f-a7d39ce5c07d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id4b49a0d637867422445508d5347169613ae1ccd