Revert "Add a new profman flag --force-merge-and-analyze."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dc352553-0fde-4a1b-949f-a7d39ce5c07d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife53d0e0ec03409f6bd0d42ec5061f73b9b18dbe