Revert "Add a new profman flag --force-merge-and-analyze."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/981f0550-96a0-4412-8156-c321815dc14c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I990abbb13277f2bfb71a3c46e904acddafd90126