Revert "Increase the hotness lookup threshold."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f9fd510d-f7dd-43fd-9978-5907ebebaea9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia489082505794f54a44f163e6d92061a90b7fad3