Revert "Put branch profiling under a flag."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de71a738-894d-484d-9608-047a6fd9c5be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I724a08b4e54950632b404b18a51a3195b8fc7fcd