Record method info on class load to simplify trace entry processing
We used to have a pass on the mutator thread to record information about
methods before doing the rest of the processing on the background
thread. This CL removes the main thread pass by recording the
information when loading any new class. We record all the methods when
starting tracing to record any existing methods. Though this by itself
doesn't improve performance, it lets us do more optimizations that would
help improve performance. This is also cleaner than the earlier
approach.

Bug: 259258187
Test: art/test.py -t 2246-trace-stream

Change-Id: Ic05a886e201927c2a1a60691ea1705c7e40ba008