[DON'T MERGE] Combined page size agnostic patches for performance measurements
This includes 2 patches:
 - Use runtime page size for large object alignment
 - Optimize access to gPageSize

Additionally, it removes the check for ctx.Config().NoBionicPageSizeMacro()
when defining the ART_PAGE_SIZE_AGNOSTIC, to enable testing of the page
size agnostic build independently of the build system configuration.

Use runtime page size for large object alignment

In general, this turns `kLargeObjectAlignment` to `gPageSize`. It avoids
consuming extra memory to align up large object to `kMaxPageSize` (16K),
when the runtime page size is 4K.

However, in order to keep `SpaceBitmap<kAlignment>` optimized, this
keeps the internal bitmap alignment fixed to the minimal page size (4K),
We then accept that if the page size is 16K at runtime, then only 1 bit
out of every 4 bits will actually be in-use.

Assuming a heap capacity of 256M, the size of the bitmap is as follows:

  - 4K:  (256M / 4K)  -> 64K pages -> 8K needed to store one bit per
    page.
  - 16K: (256M / 16K) -> 16K pages -> 2K needed to store one bit per
    page.

It seems acceptable to consume 8K regardless of the runtime page
size. We may revisit this later if we realize this is worth optimizing,
maybe more so for speed rather than memory.

To cover testing, this CL turns the existing space bitmap unit tests
into type-paramaterized tests, based on the SpaceBitmap implementation
and the object alignment they can support:

  | SpaceBitmapType       | SpaceBitmap alignment | Object alignment   |
  |-----------------------+-----------------------+--------------------|
  | ContinuousSpaceBitmap | kObjectAlignment      | kObjectAlignment   |
  | LargeObjectBimap      | kMinPageSize (4K)     | kMinPageSize (4K)  |
  | LargeObjectBimap      | kMinPageSize (4K)     | kMaxPageSize (16K) |

In passing, this also removes the templated version of
MapAnonymousAligned(), as it was only used to allocate large objects,
outside of tests. And highlights that its alignment must be higher than
gPageSize. The image tests are also adapted accordingly.

Test: art/test/testrunner/testrunner.py -b --host
  on host
Test: art/test/testrunner/testrunner.py --target --64 \
        --gcverify --gcstress
Test: art/tools/run-gtests.sh -- -j6
Test: art/tools/run-gtests.sh \
        /apex/com.android.art/bin/art/arm64/art_runtime_tests \
        -- --gtest_filter="*SpaceBitmap*"
  on target Pixel6

Change-Id: I8de16f383d6bce4a098b17f5aa328ed3d91ab16a

Optimize access to gPageSize

This change refactors the global constants dependent on gPageSize in
a way that gPageSize and dependent constants are now internal to the
runtime. Also it removes static initialization dependency on gPageSize
and disallows accessing gPageSize during the static initialization
phase. Consequently, this removes a need in the condition check when
accessing gPageSize and allows compiler to recognise the value as
a constant that can be cached where used repeatedly.

From this point, gPageSize should only be used within libart. For most
of the other cases MemMap::GetPageSize() should be used. Where MemMap
is unavailable e.g. during static initialization or another stage when
MemMap isn't yet initialized, or in a component which might operate
without MemMap being initialized, the GetPageSizeSlow() would be
generally suitable. For performance-sensitive code, GetPageSizeSlow()
shouldn't be used without caching the value to remove repeated calls
of the function.

-  The gPageSize constant is now local to libart. Access to the variable
   is disallowed until Runtime is initialized, so removing the chance of
   static initialization order issue related to this value.

   The initializers for the other constants with initialization
   that used to depend on the gPageSize value are now referring to
   GetPageSizeSlow instead, so removing the static initialization
   dependency, and consequently, possible order issues.

   This removes the condition check upon each load of the gPageSize.

   Also as the variable is defined with "hidden" visibility, it isn't
   exported and so is unavailable to other DSOs.

   This also removes one indirection level in accessing it as GOT lookup
   is not required anymore.

 - Remove gPMDSize and gPUDSize from libartbase

   Move the constants to gc::Heap in form of functions computing their
   values. Also move the BestPageTableAlignment function dependent on
   the values, as it is only used within the GC.

   Direct accesses to gPMDSize are only performed for assertions and so
   replacing that with a function call shouldn't matter for performance.

   gPUDSize isn't accessed directly.

   The rest the values accesses are done via BestPageTableAlignment,
   which is called from:
    - MarkCompact constructor
    - MarkCompact::AddLinearAllocSpaceData
    - GcVisitedArenaPool::AddMap

   In all cases, it's performed in conjunction with MapAnonymousAligned,
   which involves an mmap system call. The potential slight slowdown
   in BestPageTableAlignment - due to the dynamic computation of the
   gPMDSize and gPUDSize values - is expected to be negligible in
   comparison to the performance cost of mmap system call.

 - Move gPageSize declaration and definition to the runtime / component

   Replace the removed uses of gPageSize outside libart with
   MemMap::GetPageSize() / GetPageSizeSlow().

   There is one exception to the above statement in this change - in
   FdFile, where the page size appears to be used arbitrarily to compute
   the maximum buffer size: in that case replace the maximum buffer size
   with a constant which would have matched the value when the page size
   is 4K.

 - Make gNumLrtSlots and gStackOverflowProtectedSize member variables

   The gNumLrtSlots and gStackOverflowProtectedSize globals are only
   used within their respective modules and only within
   the corresponding classes:
    - SmallLrtAllocator for gNumLrtSlots
    - Thread for gStackOverflowProtectedSize

   Move them inside the corresponding classes as member variables.

 - Remove ART_PAGE_SIZE_AGNOSTIC_DECLARE_* macros
   The macros aren't used anymore. Remove them and GlobalConst.

The tests were run for legacy 4K, page size agnostic 4K and 16K.

Co-authored-by: Ruben Ayrapetyan <ruben.ayrapetyan@arm.com>
Co-authored-by: Richard Neill <richard.neill@arm.com>
Test: art/tools/run-gtests.sh
Test: art/test/testrunner/testrunner.py --target --64
Test: art/tools/run-libcore-tests.sh --mode=device --variant=X64
Test: art/tools/run-libjdwp-tests.sh --mode=device --variant=X64
Test: cd art && mma
Change-Id: I07fd85cd35dd443dee5cff0f7435ab47ba727c6f

Change-Id: I4d60434fff374a723e8272a9465d396b31445260